[MODIFY] UserResponseDto 생성하여 회원과 관련된 DTO 관리
