+++ src/main/java/com/dnd/ground/domain/user/dto/HomeResponseDto.java
@@ -24,44 +24,12 @@ import java.util.*;
 public class HomeResponseDto {
 
     @ApiModelProperty(value="유저에 대한 정보", required = true)
-    private UserMatrix userMatrices;
+    private UserResponseDto.UserMatrix userMatrices;
 
     @ApiModelProperty(value="(챌린지를 안하는)친구들에 대한 정보", required = true)
-    private List<FriendMatrix> friendMatrices;
+    private List<UserResponseDto.FriendMatrix> friendMatrices;
 
     @ApiModelProperty(value="(챌린지를 하는)유저들에 대한 정보",  required = true)
-    private List<ChallengeMatrix> challengeMatrices;
+    private List<UserResponseDto.ChallengeMatrix> challengeMatrices;
 
-    @AllArgsConstructor
-    static public class UserMatrix{
-        @ApiModelProperty(value = "닉네임", example = "NickA", required = true)
-        public String nickname;
-
-        @ApiModelProperty(value = "칸 꼭지점 위도, 경도 리스트", required = true)
-        public Set<MatrixSetDto> matrices;
-    }
-
-    @AllArgsConstructor
-    static public class FriendMatrix{
-        @ApiModelProperty(value = "닉네임", example = "NickB", required = true)
-        public String nickname;
-
-        @ApiModelProperty(value = "칸 꼭지점 위도, 경도 리스트", required = true)
-        public Set<MatrixSetDto> matrices;
-    }
-
-    @AllArgsConstructor
-    static public class ChallengeMatrix{
-        @ApiModelProperty(value = "닉네임", example = "NickC", required = true)
-        public String nickname;
-
-        @ApiModelProperty(value = "나와 같이 하는 챌린지 개수", example = "1", required = true)
-        public Integer challengeNumber;
-
-        @ApiModelProperty(value = "지도에 나타나는 챌린지 대표 색깔", example = "#ffffff", required = true)
-        public String challengeColor;
-
-        @ApiModelProperty(value = "칸 꼭지점 위도, 경도 리스트", required = true)
-        public Set<MatrixSetDto> matrices;
-    }
 }
